+++ packages/card-builder/src/exportApi.ts
@@ -1,4 +1,4 @@
-import type { CardConfig } from "./export-helpers";
+import type { CardConfig } from "./export-utils";
 
 // Convert a CardConfig into an OpenAPI 3.0 YAML document.
 // Each interactive element gets a minimal POST endpoint so consumers
